Split `fb-harmony` into `flow` and `jsx` files.
Closes #138.
