+++ main.js
@@ -10,7 +10,8 @@ require("./def/es6");
 require("./def/es7");
 require("./def/mozilla");
 require("./def/e4x");
-require("./def/fb-harmony");
+require("./def/jsx");
+require("./def/flow");
 require("./def/esprima");
 require("./def/babel");
